refactor: created a common component for dataset operation menu
